This is a breaking change. I should add semver support.
This is a breaking change to this repo. It is not backwards compatible.
CMake is now used to build the example. First time using it, its probably wrong.
Major change is that initializing the cli now requires you pass a buffer. This allows the user to declare the buffer used without dynamic allocation of memory and moves away from the static, fixed memory allocation. This allows the cli to be more configurable per target.
Tests updated. Added support to make sure we can support a very large buffer - this is the reason for this change.
I'm still not a huge fan that the library handles the output. I also figured out how to make this more testable via the print callback. Need to update the remaining tests to verify the output.